+++ lib/data_monitor_generators.dart
@@ -5,7 +5,7 @@ import 'package:source_gen/source_gen.dart';
 
 import 'package:data_monitor_generators/src/data_monitor_generator.dart';
 
-/// Called by 'build_runner' to generate required '.g.dart' files.
+/// Called by `build_runner` to generate required '.g.dart' files.
 Builder generateBuilders(BuilderOptions options) {
   return SharedPartBuilder([DataMonitorGenerator()], 'data_monitor');
 }