+++ lib/src/data_monitor_generator.dart
@@ -5,32 +5,59 @@ import 'package:source_gen/source_gen.dart';
 
 import 'package:data_monitor/annotations.dart';
 
-import 'model_visitor.dart';
+import 'class_visitor.dart';
 
 /// `DataMonitorGenerator` uses the information that was found
-/// when any `data_monitor` related annotations were parsed by
+/// when any classes annotated by @dataMonitor were parsed by
 /// a generator.
 ///
 /// It creates concrete classes in (multiple) '.g.dart' files
-/// that implement the required field `setters` in any
-/// associated data model classes.
+/// that implement the required `field setters` and `property
+/// getters` in in any associated data model classes.
 class DataMonitorGenerator extends GeneratorForAnnotation<DataMonitor> {
   @override
-  String generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep) {
-    final ModelVisitor visitor = ModelVisitor();
-    final StringBuffer buffer  = StringBuffer();
+  Future<String> generateForAnnotatedElement(Element element, ConstantReader annotation, BuildStep buildStep) async {
+    final ClassVisitor          visitor    = ClassVisitor(resolver: buildStep.resolver);
+    final Map<String, DartType> properties = {};
+    final StringBuffer          buffer     = StringBuffer();
 
-    element.visitChildren(visitor);
+    if (element is ClassElement) {
+      await visitor.visitClass(element);
 
-    buffer.writeln('mixin _\$${visitor.className} on ${visitor.className} {');
+      buffer.writeln('mixin _\$${visitor.className} on ${visitor.className} {');
 
-    for (MapEntry<String, DartType> field in visitor.fields.entries) {
-      buffer.writeln('@override');
-      buffer.writeln('set ${field.key}(${field.value.toString()} value) {');
-      buffer.writeln('  notify(() {');
-      buffer.writeln('    super.${field.key} = value;');
-      buffer.writeln('  });');
-      buffer.writeln('}');
+      for (MapEntry<String, FieldContainer> field in visitor.fields.entries) {
+        String name = field.key;
+
+        properties.addAll(field.value.properties);
+
+        buffer.writeln('late final DataItem _${field.key} = DataItem(${field.key}, [');
+
+        for (String name in field.value.properties.keys) {
+          buffer.writeln('  _$name,');
+        }
+        buffer.writeln(']);');
+
+        buffer.writeln('');
+        buffer.writeln('@override');
+        buffer.writeln('set $name(${field.value.type} value) {');
+        buffer.writeln('  if (_$name.setValue(value)) {');
+        buffer.writeln('    notify(() {');
+        buffer.writeln('      super.$name = value;');
+        buffer.writeln('    });');
+        buffer.writeln('  }');
+        buffer.writeln('}');
+      }
+      for (MapEntry<String, DartType> property in properties.entries) {
+        String name = property.key;
+
+        buffer.writeln('late final DataProperty _$name = DataProperty(super.$name);');
+        buffer.writeln('');
+        buffer.writeln('@override');
+        buffer.writeln('${property.value} get $name {');
+        buffer.writeln('  return _$name.isDirty ? _$name.setValue(super.$name) : _$name.value;');
+        buffer.writeln('}');
+      }
     }
     buffer.writeln('}');