+++ CHANGELOG.md
@@ -1,5 +1,23 @@
+## 1.1.3
+
+Now emits correct code in '.g.dart' files to cater for new
+`@cache` annotation using `DataItem` and `DataProperty`
+classes.
+
+Big asynchronicity issue fixed in main generator.
+
 ## 1.1.2
 
 With modified example.md
 
 Matches data_monitor version v1.1.2
+
+## 1.1.1
+
+With modified example.md
+
+Matches data_monitor version v1.1.0
+
+## 1.1.0
+
+Removed beta release warning and promoted to v1.1.0